+++ src/main/java/blackjack/domain/CardDeck.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
 import java.util.ArrayDeque;
 import java.util.Arrays;
 import java.util.Collections;
