+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,9 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static blackjack.domain.GameOutcome.DRAW;
 import static blackjack.domain.GameOutcome.LOSE;
 import static blackjack.domain.GameOutcome.WIN;
 
+import blackjack.domain.GameOutcome;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
